added column of core_complex ids / Keith's pr suggestions
Former-commit-id: 8edf78a9c004482522edcca2f1a0995a07ac0ec7

+++ opencell/mass_spec/clusterone_utils.py
@@ -443,7 +443,7 @@ def cluster_matrix_to_sql_table(mcl, network, method='single', metric='cosine', 
     sql['row_index'] = rows_col
     sql['pulldown_plate_id'] = pull_plate_col
 
-    sql['subcluster_id'] = sql['subcluster_id'].astype('Int64')
+    # sql['subcluster_id'] = sql['subcluster_id'].astype('Int64')
     return sql
